HERITAGE-135: Filters (#50)
* HERITAGE-135:move prepare filter agg to utils

* HERITAGE-135:remove API OR operator for collection filter
